# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend for the all rated extreme demon list",
        "title": "Aredl API",
        "contact": {
            "name": "Aredl",
            "url": "https://aredl.net/"
        },
        "version": "1.0"
    },
    "host": "api.aredl.net",
    "basePath": "/api",
    "paths": {
...
        "/aredl/me/submissions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Lists submissions ordered by the time they have been updated last.\nRequires user permission: aredl.user_submission_list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aredl"
                ],
                "summary": "List submissions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/aredl.MeSubmission"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a submission. If a submission for a level already exists, it will be updated instead. Submissions can only be updated when its status is pending or rejected_retryable\nRequires user permission: aredl.user_submit\nIf the user has the permission aredl.priority they will automatically be assigned to the priority queue",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aredl"
                ],
                "summary": "Create or Update submission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "internal level id",
                        "name": "level",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "url",
                        "description": "display video url",
                        "name": "video_url",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "whether submission was done on mobile",
                        "name": "mobile",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ldm gd level id if used",
                        "name": "ldm_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "url",
                        "description": "raw footage",
                        "name": "raw_footage",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "additional notes the user wants to add to a submission. Max 100 characters",
                        "name": "additional_notes",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/aredl/me/submissions/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes a submission as long as it still is open for review.\nRequires user permission: aredl.user_submission_delete",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aredl"
                ],
                "summary": "Delete submission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "submission id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/aredl/packs": {
            "get": {
                "description": "Gives a list of all packs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aredl"
                ],
                "summary": "Aredl packs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/aredl.Pack"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new pack and updates all user points that now have the new pack.\nRequires user permission: aredl.manage_packs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aredl"
                ],
                "summary": "Create a new AREDL pack",
                "parameters": [
                    {
                        "type": "string",
                        "description": "display name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "display color",
                        "name": "color",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "position of pack",
                        "name": "placement_order",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "list of internal level ids. Pack has to have at least two levels",
                        "name": "levels",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "aredl.MeSubmission": {
            "type": "object",
            "properties": {
                "additional_notes": {
                    "type": "string"
                },
                "created": {
                    "$ref": "#/definitions/types.DateTime"
                },
                "id": {
                    "type": "string"
                },
                "is_update": {
                    "type": "boolean"
                },
                "ldm_id": {
                    "type": "integer"
                },
                "level": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "string"
                        },
                        "level_id": {
                            "type": "integer"
                        },
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "mobile": {
                    "type": "boolean"
                },
                "priority": {
                    "type": "boolean"
                },
                "raw_footage": {
                    "type": "string"
                },
                "rejected": {
                    "type": "boolean"
                },
                "updated": {
                    "$ref": "#/definitions/types.DateTime"
                },
                "video_url": {
                    "type": "string"
                }
            }
        },
...
        "aredl.Pack": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "levels": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "legacy": {
                                "type": "boolean"
                            },
                            "level_id": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            },
                            "points": {
                                "type": "number"
                            },
                            "position": {
                                "type": "integer"
                            }
                        }
                    }
                },
                "name": {
                    "type": "string"
                },
                "points": {
                    "type": "number"
                }
            }
        },
...
        "types.DateTime": {
            "type": "object"
        },
        "util.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Perform actions as a user. It is also used to access endpoints that require user permissions.",
            "type": "apiKey",
            "name": "api-key",
            "in": "header"
        }
    }
}